Add complete dependency list on Ubuntu
In order to figure this out (and make sure I was correct) I used a Docker container. By design it's awkward to support dbus & bluetooth from within a container, so totally understandable if you'd prefer to not include the Dockerfile in the repo.

I managed to get the app to run and print out "Bluetooth is disabled" which exercises a decent bit of the code. If this is helpful, it could be the start of some light CI/GitHub workflows to run some testing for PRs.