sync persist-pair with what's in main (#76)
* Add complete dependency list on Ubuntu (#55)

In order to figure this out (and make sure I was correct) I used a Docker container. By design it's awkward to support dbus & bluetooth from within a container, so totally understandable if you'd prefer to not include the Dockerfile in the repo.

I managed to get the app to run and print out "Bluetooth is disabled" which exercises a decent bit of the code. If this is helpful, it could be the start of some light CI/GitHub workflows to run some testing for PRs.

Co-authored-by: Alex R <m.alexanderrobinson@yahoo.com>

* change app-id to com.github.alexr4535.siglo for flathub

* update manifest for flathub

* bump tag

* Update com.github.alexr4535.siglo.json

* add python3-modules.json

* reference python3-requests in manifest

* prepare for next release

* use --system-talk-name to specify dbus name

* Install the required Python packages in the Flatpak manifest (#61)

* Prepare for next release

* Appdata: Add missing XML tags required by flathub

* Appdata: Add screenshot required by flathub

* bump version for next release

* fix typo

* bump version for next release

* bump version for next release

* Appdata: fix screenshot URL required by flathub

* bump version for next release

* Adding custom form_factor values (#67)

This should make Siglo show up in the PureOS Store on the Librem 5

* Adding X-Purism-FormFactor (#66)

This should help with showing up on the main app launcher of Phosh > 0.12.
For details read https://linmob.net/phosh-0-12-app-drawer/

* Fix metadata file (#68)

Fixes https://github.com/alexr4535/siglo/issues/65

* Bump version for next release

* Fix crash without network access (#69)

This allows Siglo to start when internet access is not available, either
because the user is outside service or the flatpak permission has been
disabled.

Co-authored-by: undef <gitlab@undef.tools>

* Add requirements section and flathub install instructions

* Version Bump

Co-authored-by: Joe Smith <yasumoto7@gmail.com>
Co-authored-by: Jordan Williams <jordan@jwillikers.com>
Co-authored-by: 1peter10 <32742559+1peter10@users.noreply.github.com>
Co-authored-by: TheEvilSkeleton <theevilskeleton@riseup.net>
Co-authored-by: Undef-a <75462734+Undef-a@users.noreply.github.com>
Co-authored-by: undef <gitlab@undef.tools>