Allow a name suffix in the bluetooth alias (#90)
Currently PineTime watches are detected by the device alias so if the
user changes the alias they are no longer listed in Siglo. This is
pretty hard to fix because there's no good way to actually detect it any
other way. This fix lets people at least add an suffix to the alias to
differentiate devices.