Manually get `XDG_CONFIG_HOME` and drop pyxdg (#102)
Drop `pyxdg` as dependency and instead get `XDG_CONFIG_HOME` manually.
The same way is being used as `pyxdg` does:
https://gitlab.freedesktop.org/xdg/pyxdg/-/blob/bd999c1c/xdg/BaseDirectory.py#L37-L38

There is a name clash between the two packages `xdg` and `pyxdg`. See:
https://gitlab.freedesktop.org/xdg/pyxdg/-/issues/24

As only that one environment variable is being used from the library
it can just as good be expanded manually.
This solves the confusion and problems for users who have both
installed and it also reduces the dependency footprint.

Also use `os.path.join` for `config_file`.